Move abstract Object.assign temporals into a helper funciton
